03_basics_gcf/python/requirements.txt
google-cloud-iot